Enum mentat_cli::command_parser::Command

pub enum Command {
    Cache(String, CacheDirection),
    Close,
    Exit,
    Help(Vec<String>),
    Import(String),
    Open(String),
    OpenEmpty(String),
    OpenEncrypted(String, String),
    OpenEmptyEncrypted(String, String),
    Query(String),
    QueryExplain(String),
    QueryPrepared(String),
    Schema,
    Sync(Vec<String>),
    Timer(bool),
    Transact(String),
}

Variants
Cache(String, CacheDirection)CloseExitHelp(Vec<String>)Import(String)Open(String)OpenEmpty(String)OpenEncrypted(String, String)OpenEmptyEncrypted(String, String)Query(String)QueryExplain(String)QueryPrepared(String)SchemaSync(Vec<String>)Timer(bool)Transact(String)
Methods

impl Command

pub fn is_complete(&self) -> bool

is_complete returns true if no more input is required for the command to be successfully executed. false is returned if the command is not considered valid. Defaults to true for all commands except Query and Transact. TODO: for query and transact commands, they will be considered complete if a parsable EDN has been entered as an argument

pub fn is_timed(&self) -> bool

pub fn output(&self) -> String